requirements.txt
xgboost==2.0.3
evidently==0.4.30
fastapi==0.110.0
uvicorn==0.29.0
streamlit==1.40.0
pandas==2.2.1
scikit-learn==1.4.1.post1
python-dotenv==1.0.1
pydantic==2.6.4
plotly==5.20.0
numpy==1.26.4
